// src/common/enums/http-codes.enum.ts
export enum HttpCodesEnum {
  HTTP_OK = 200,
  HTTP_CREATED = 201,

  HTTP_BAD_REQUEST = 400,
  HTTP_UNAUTHORIZED = 401,
  HTTP_FORBIDDEN = 403,
  HTTP_NOT_FOUND = 404,
  HTTP_CONFLICT = 409,
  HTTP_UNPROCESSABLE_ENTITY = 422,
  HTTP_TOO_MANY_REQUESTS = 429,

  HTTP_SERVER_ERROR = 500,
  HTTP_SERVICE_UNAVAILABLE = 503,
}